[refactor] Move inconsistency check functions to a new util function file (#3866)
