+++ ray-operator/controllers/ray/raycluster_controller_unit_test.go
@@ -1606,146 +1606,6 @@ func TestReconcile_UpdateClusterState(t *testing.T) {
 	assert.Equal(t, cluster.Status.State, state, "Cluster state should be updated")
 }
 
-func TestInconsistentRayClusterStatus(t *testing.T) {
-	newScheme := runtime.NewScheme()
-	_ = rayv1.AddToScheme(newScheme)
-	fakeClient := clientFake.NewClientBuilder().WithScheme(newScheme).WithRuntimeObjects().Build()
-	r := &RayClusterReconciler{
-		Client:   fakeClient,
-		Recorder: &record.FakeRecorder{},
-		Scheme:   scheme.Scheme,
-	}
-
-	// Mock data
-	timeNow := metav1.Now()
-	oldStatus := rayv1.RayClusterStatus{
-		State:                   rayv1.Ready,
-		ReadyWorkerReplicas:     1,
-		AvailableWorkerReplicas: 1,
-		DesiredWorkerReplicas:   1,
-		MinWorkerReplicas:       1,
-		MaxWorkerReplicas:       10,
-		LastUpdateTime:          &timeNow,
-		Endpoints: map[string]string{
-			utils.ClientPortName:    strconv.Itoa(utils.DefaultClientPort),
-			utils.DashboardPortName: strconv.Itoa(utils.DefaultDashboardPort),
-			utils.GcsServerPortName: strconv.Itoa(utils.DefaultGcsServerPort),
-			utils.MetricsPortName:   strconv.Itoa(utils.DefaultMetricsPort),
-		},
-		Head: rayv1.HeadInfo{
-			PodIP:     "10.244.0.6",
-			ServiceIP: "10.96.140.249",
-		},
-		ObservedGeneration: 1,
-		Reason:             "test reason",
-	}
-
-	// `inconsistentRayClusterStatus` is used to check whether the old and new RayClusterStatus are inconsistent
-	// by comparing different fields. If the only differences between the old and new status are the `LastUpdateTime`
-	// and `ObservedGeneration` fields, the status update will not be triggered.
-	ctx := context.Background()
-
-	testCases := []struct {
-		modifyStatus func(*rayv1.RayClusterStatus)
-		name         string
-		expectResult bool
-	}{
-		{
-			name: "State is updated, expect result to be true",
-			modifyStatus: func(newStatus *rayv1.RayClusterStatus) {
-				newStatus.State = rayv1.Suspended
-			},
-			expectResult: true,
-		},
-		{
-			name: "Reason is updated, expect result to be true",
-			modifyStatus: func(newStatus *rayv1.RayClusterStatus) {
-				newStatus.Reason = "new reason"
-			},
-			expectResult: true,
-		},
-		{
-			name: "ReadyWorkerReplicas is updated, expect result to be true",
-			modifyStatus: func(newStatus *rayv1.RayClusterStatus) {
-				newStatus.ReadyWorkerReplicas = oldStatus.ReadyWorkerReplicas + 1
-			},
-			expectResult: true,
-		},
-		{
-			name: "AvailableWorkerReplicas is updated, expect result to be true",
-			modifyStatus: func(newStatus *rayv1.RayClusterStatus) {
-				newStatus.AvailableWorkerReplicas = oldStatus.AvailableWorkerReplicas + 1
-			},
-			expectResult: true,
-		},
-		{
-			name: "DesiredWorkerReplicas is updated, expect result to be true",
-			modifyStatus: func(newStatus *rayv1.RayClusterStatus) {
-				newStatus.DesiredWorkerReplicas = oldStatus.DesiredWorkerReplicas + 1
-			},
-			expectResult: true,
-		},
-		{
-			name: "MinWorkerReplicas is updated, expect result to be true",
-			modifyStatus: func(newStatus *rayv1.RayClusterStatus) {
-				newStatus.MinWorkerReplicas = oldStatus.MinWorkerReplicas + 1
-			},
-			expectResult: true,
-		},
-		{
-			name: "MaxWorkerReplicas is updated, expect result to be true",
-			modifyStatus: func(newStatus *rayv1.RayClusterStatus) {
-				newStatus.MaxWorkerReplicas = oldStatus.MaxWorkerReplicas + 1
-			},
-			expectResult: true,
-		},
-		{
-			name: "Endpoints is updated, expect result to be true",
-			modifyStatus: func(newStatus *rayv1.RayClusterStatus) {
-				newStatus.Endpoints["fakeEndpoint"] = "10009"
-			},
-			expectResult: true,
-		},
-		{
-			name: "Head.PodIP is updated, expect result to be true",
-			modifyStatus: func(newStatus *rayv1.RayClusterStatus) {
-				newStatus.Head.PodIP = "test head pod ip"
-			},
-			expectResult: true,
-		},
-		{
-			name: "RayClusterReplicaFailure is updated, expect result to be true",
-			modifyStatus: func(newStatus *rayv1.RayClusterStatus) {
-				meta.SetStatusCondition(&newStatus.Conditions, metav1.Condition{Type: string(rayv1.RayClusterReplicaFailure), Status: metav1.ConditionTrue})
-			},
-			expectResult: true,
-		},
-		{
-			name: "LastUpdateTime is updated, expect result to be false",
-			modifyStatus: func(newStatus *rayv1.RayClusterStatus) {
-				newStatus.LastUpdateTime = &metav1.Time{Time: timeNow.Add(time.Hour)}
-			},
-			expectResult: false,
-		},
-		{
-			name: "ObservedGeneration is updated, expect result to be false",
-			modifyStatus: func(newStatus *rayv1.RayClusterStatus) {
-				newStatus.ObservedGeneration = oldStatus.ObservedGeneration + 1
-			},
-			expectResult: false,
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			newStatus := oldStatus.DeepCopy()
-			testCase.modifyStatus(newStatus)
-			result := r.inconsistentRayClusterStatus(ctx, oldStatus, *newStatus)
-			assert.Equal(t, testCase.expectResult, result)
-		})
-	}
-}
-
 func TestCalculateStatus(t *testing.T) {
 	setupTest(t)
 	assert.True(t, features.Enabled(features.RayClusterStatusConditions))
