+++ ray-operator/controllers/ray/rayservice_controller_unit_test.go
@@ -61,52 +61,6 @@ func TestGenerateHashWithoutReplicasAndWorkersToDelete(t *testing.T) {
 	assert.NotEqual(t, hash1, hash3)
 }
 
-func TestInconsistentRayServiceStatuses(t *testing.T) {
-	oldStatus := rayv1.RayServiceStatuses{
-		ActiveServiceStatus: rayv1.RayServiceStatus{
-			RayClusterName: "new-cluster",
-			Applications: map[string]rayv1.AppStatus{
-				utils.DefaultServeAppName: {
-					Status:  rayv1.ApplicationStatusEnum.RUNNING,
-					Message: "OK",
-					Deployments: map[string]rayv1.ServeDeploymentStatus{
-						"serve-1": {
-							Status:  rayv1.DeploymentStatusEnum.UNHEALTHY,
-							Message: "error",
-						},
-					},
-				},
-			},
-		},
-		PendingServiceStatus: rayv1.RayServiceStatus{
-			RayClusterName: "old-cluster",
-			Applications: map[string]rayv1.AppStatus{
-				utils.DefaultServeAppName: {
-					Status:  rayv1.ApplicationStatusEnum.NOT_STARTED,
-					Message: "application not started yet",
-					Deployments: map[string]rayv1.ServeDeploymentStatus{
-						"serve-1": {
-							Status:  rayv1.DeploymentStatusEnum.HEALTHY,
-							Message: "Serve is healthy",
-						},
-					},
-				},
-			},
-		},
-		ServiceStatus: rayv1.NotRunning,
-	}
-	ctx := context.Background()
-
-	// Test 1: Update ServiceStatus only.
-	newStatus := oldStatus.DeepCopy()
-	newStatus.ServiceStatus = rayv1.Running
-	assert.True(t, inconsistentRayServiceStatuses(ctx, oldStatus, *newStatus))
-
-	// Test 2: Test RayServiceStatus
-	newStatus = oldStatus.DeepCopy()
-	assert.False(t, inconsistentRayServiceStatuses(ctx, oldStatus, *newStatus))
-}
-
 func TestIsHeadPodRunningAndReady(t *testing.T) {
 	// Create a new scheme with CRDs, Pod, Service schemes.
 	newScheme := runtime.NewScheme()
